% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{alt_bav_tk4tel}
\alias{alt_bav_tk4tel}
\title{altitudinal and terrain data of bavaria}
\format{
A \code{data.frame} with 2268 observations and 11 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://gdex.cr.usgs.gov/gdex/}}
}
\usage{
data(alt_bav_tk4tel)
}
\description{
NASA SRTM v3.0 elevation data of Bavaria, Germany
}
\details{
Elevation and terrain data of Bavaria, Germany derived from NASA SRTM 3 Arc Sec v3.0 data and 
resampled onto the TK4tel grid. Terrain variables were calculated using the raster::terrain() function. 
Data includes information on altitude, aspect, slope, hillshade (calculated from slope and aspect), 
Terrain Ruggedness Index (tri), Topographic Position Index (tpi), 
Topographic Position Index with neighborhood size 5 (tpi5), roughness and flow direction (flowdir).
}
