% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{amphibians_bav}
\alias{amphibians_bav}
\title{amphibian range maps for the extent of Bavaria}
\format{A \code{data.frame} with 23 observations and 26 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://www.iucnredlist.org/resources/spatial-data-download}}
}
\usage{
data(amphibians_bav)
}
\description{
IUCN Amphibian ranges with Bavaria, Germany
}
\details{
shapefiles of the global IUCN ranges of all amphibian species that intersect with the extent of Bavaria
}
