% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{bav_pop_1km}
\alias{bav_pop_1km}
\title{Population density of Bavaria}
\format{
A \code{data.frame} with 70552 observations and 5 variables.
}
\usage{
data(bav_pop_1km)
}
\description{
Population density in Bavaria, Germany
}
\details{
Population density of Bavaria at a resolution of 1km2
}
\references{
Schug, F., Frantz, D., van der Linden, S., & Hostert, P. (2021). 
 Gridded population mapping for Germany based on building density, height and type from 
 Earth Observation data using census disaggregation and bottom-up estimates. 
 PLOS ONE. \url{10.1371/journal.pone.0249044}
}
