% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{bavaria}
\alias{bavaria}
\title{administrative boundary of bavaria}
\format{A \code{data.frame} with 1 observation and 13 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://www.gadm.org/data.html}}
}
\usage{
data(bavaria)
}
\description{
GADM outline of Bavaria, Germany
}
\details{
Administrative boundary of Bavaria, Germany derived from GADM data.
}
