% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_flag_transposed_xy.R
\name{bcd_flag_transposed_xy}
\alias{bcd_flag_transposed_xy}
\title{Flag transposed coordinates}
\usage{
bcd_flag_transposed_xy(data)
}
\arguments{
\item{data}{a data.frame with the default column names: "database_id", "scientificName", "decimalLongitude", "decimalLatitude"}
}
\description{
This function add a new column \code{.transposed_xy} in the returned dataset
}
\examples{
\dontrun{
data \%>\%
  bdc_flag_transposed_xy()
}
}
