% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_aux_functions.R
\name{bdc_check_flags}
\alias{bdc_check_flags}
\title{Title}
\usage{
bdc_check_flags(data)
}
\arguments{
\item{raw_data}{}

\item{filtered_data}{}

\item{save_in_filename}{}

\item{comment}{}
}
\value{

}
\description{
Title
}
