% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_coord_trans.R
\name{bdc_coord_trans}
\alias{bdc_coord_trans}
\title{Title: Corrected inverted xy coordinates}
\usage{
bdc_coord_trans(data, x, y, country_code, id, worldmap, worldmap_cntr_code)
}
\arguments{
\item{data}{}

\item{x}{}

\item{y}{}

\item{country_code}{}

\item{id}{}

\item{worldmap}{}

\item{worldmap_cntr_code}{}
}
\value{

}
\description{
Title: Corrected inverted xy coordinates
}
