% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_coordinates_country_inconsistent.R
\name{bdc_coordinates_country_inconsistent}
\alias{bdc_coordinates_country_inconsistent}
\title{Identify records within a reference country
This function flags geographic coordinates within a reference country. A spatial buffer can be added to the reference country for ensuring that records in mangroves, marshes, estuaries as well as records with low coordinates precision are flagged as invalid.}
\usage{
bdc_coordinates_country_inconsistent(
  data,
  country_name,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  dist = 0.1
)
}
\arguments{
\item{data}{data.frame. Containing longitude and latitude. Coordinates must
be expressed in decimal degree and in WGS84.}

\item{country_name}{character string. Name of the country to considered.}

\item{lat}{character string. The column name with the latitude coordinates.
Default = “decimallatitude”.}

\item{lon}{character string. The column name with the longitude coordinates.
Default = “decimallongitude”.}

\item{dist}{numeric. A distance in decimal degrees used to created a buffer
around the country. Default = 0.1 (~11 km at the equator)}
}
\value{
A data.frame contain the column '.coordinates_country_inconsistent'.
Compliant (TRUE) if coordinates fall within the boundaries plus a specified distance (if 'dist' is supplied) of 'country_name'; otherwise "FALSE".
}
\description{
Records in other countries or in the ocean at a specified distance
from the coastline of the reference country, which
}
\details{
The distance informed in the argument 'dist' is used to create a
buffer around the reference country. Records within the reference country
or at a specified distance from the coastline of the reference country
(i.e., records within the buffer) are flagged as valid (TRUE). Note that
records within the buffer but in other countries are flagged as invalid
(FALSE).
}
\examples{
\dontrun{
decimalLatitude <- c(19.9358, -13.016667, NA, "")
decimalLongitude <- c(-40.6003, -39.6, -20.5243, NA)
x <- data.frame(decimalLatitude, decimalLongitude)

bdc_coordinates_country_inconsistent(
  data = x,
  country_name = "Brazil",
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  dist = 0.1 # in decimal degrees
)
}
}
