% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_coordinates_from_locality.R
\name{bdc_coordinates_from_locality}
\alias{bdc_coordinates_from_locality}
\title{Identify records missing or with invalid coordinates but with information on
locality}
\usage{
bdc_coordinates_from_locality(
  data,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  locality = "locality"
)
}
\arguments{
\item{data}{data.frame. Containing geographical coordinates and the column
"locality'.}

\item{lat}{character string. The column name with latitude. Coordinates must
be expressed in decimal degree and in WGS84. Default = "decimalLatitude".}

\item{lon}{character string. The column with longitude. Coordinates must be
expressed in decimal degree and in WGS84. Default = "decimalLongitude".}

\item{locality}{character string. The column name with information on
locality. Default = "locality".}
}
\value{
A data.frame containing records missing or with invalid coordinates
but with potentially useful locality information is saved in
Output/Check/01_coordinates_from_locality.csv.
}
\description{
This function Identifies records whose coordinates can potentially be
extracted from information on locality.
}
\details{
According to DarwinCore terminology, locality refers to "the
specific description of the place" where a organism were recorded.
}
\examples{
\dontrun{
lat <- c(NA, NA, "")
lon <- c("", NA, NA)
x <- data.frame(lat = lat, lon = lon, locality =  c("Brazil", "Argentina", "Chile"))
bdc_coordinates_from_locality(data = x, lat = "lat", lon = "lon")
}
}
