% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_coordinates_out_country.R
\name{bdc_coordinates_out_country}
\alias{bdc_coordinates_out_country}
\title{Identify records outside a reference country
This function flags geographic coordinates outside the reference country
(i.e., records in other countries or in the ocean at a specified distance
from the coastline of the reference country). This avoids flagging as invalid
records in mangroves, marshes, estuaries, etc.}
\usage{
bdc_coordinates_out_country(
  data,
  country_name,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  dist = 0.1
)
}
\arguments{
\item{data}{data.frame. Containing longitude and latitude. Coordinates must
be expressed in decimal degree and in WGS84.}

\item{country_name}{character string. Name of the country to considered.}

\item{lat}{character string. The column name with the latitude coordinates.
Default = “decimallatitude”.}

\item{lon}{character string. The column name with the longitude coordinates.
Default = “decimallongitude”.}

\item{dist}{numeric. A distance in decimal degrees used to created a buffer
around the country. Default = 0.1 (~11 km at the equator)}
}
\value{
A data.frame contain the column '.invalid_coordinates'. Records that
have failed in the test are flagged as "FALSE".
}
\description{
Identify records outside a reference country
This function flags geographic coordinates outside the reference country
(i.e., records in other countries or in the ocean at a specified distance
from the coastline of the reference country). This avoids flagging as invalid
records in mangroves, marshes, estuaries, etc.
}
\details{
The distance informed in the argument 'dist' is used to create a
buffer around the reference country. Records within the reference country
or at a specified distance from the coastline of the reference country
(i.e., records within the buffer) are flagged as valid (TRUE). Note that
records within the buffer but in other countries are flagged as invalid
(FALSE).
}
\examples{
\dontrun{
decimalLatitude <- c(19.9358, -13.016667, NA, "")
decimalLongitude <- c(-40.6003, -39.6, -20.5243, NA)
x <- data.frame(decimalLatitude, decimalLongitude)

bdc_coordinates_out_country(
  data = x,
  country_name = "Brazil",
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  dist = 0.1 # in decimal degrees
)
}
}
