% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_coordinates_precision.R
\name{bdc_coordinates_precision}
\alias{bdc_coordinates_precision}
\title{Flag low-precise geographic coordinates}
\usage{
bdc_coordinates_precision(
  data,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  ndec = c(0, 1, 2)
)
}
\arguments{
\item{data}{data.frame. A data.frame containing geographic coordinates.}

\item{lat}{character string. The column with latitude in decimal degrees and
WGS84. Default = "decimalLatitude".}

\item{lon}{character string. The column with longitude in decimal degrees and
WGS84. Default = "decimalLongitude".}

\item{ndec}{numeric. The numbers of decimals place to be tested. Default =
c(0,1,2).}
}
\value{
A data.frame with logical values indicating whether values are
rounded by the specified decimal number (ndec). In other words, potentially
imprecise coordinates.
}
\description{
This function is used to flag records with a coordinate precision below a
specified number of decimal places. Coordinates with one, two, or three
decimal places present a precision of~11.1 km, ~1.1 km, and ~111 m at the
equator, respectively.
}
\examples{
\dontrun{
lat = c(-21.345, 23.567, 16.798, -10.468)
lon = c(-55.389, -13.897, 30.678, 90.675)
x <- data.frame(lat, lon)

bdc_coordinates_precision(
data = x,
lat = "lat",
lon = "lon",  
ndec = c(0, 2))
}
}
