% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{bdc_correct_coordinates}
\alias{bdc_correct_coordinates}
\title{Title: correct_coordinates is a function that will detect those occurrences georreferenced outside their country different coordinate transformation}
\usage{
bdc_correct_coordinates(
  data,
  x,
  y,
  sp,
  id,
  cntr_iso2,
  world_poly,
  world_poly_iso
)
}
\arguments{
\item{data}{}

\item{x}{}

\item{y}{}

\item{sp}{}

\item{id}{}

\item{cntr_iso2}{}

\item{world_poly}{}

\item{world_poly_iso}{}
}
\value{

}
\description{
Title: correct_coordinates is a function that will detect those occurrences georreferenced outside their country different coordinate transformation
}
