% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_country_from_coordinates.R
\name{bdc_country_from_coordinates}
\alias{bdc_country_from_coordinates}
\title{Get country names from coordinates}
\usage{
bdc_country_from_coordinates(
  data,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  country = "country"
)
}
\arguments{
\item{data}{data.frame. Containing geographical coordinates and country
names. Coordinates must be expressed in decimal degree and in WGS84.}

\item{lat}{character string. The column name with latitude. Coordinates must
be expressed in decimal degree and in WGS84. Default = "decimalLatitude".}

\item{lon}{character string. The column with longitude. Coordinates must be
expressed in decimal degree and in WGS84. Default = "decimalLongitude".}

\item{country}{character string. The column name with the country assignment
of each record. Default = "country".}
}
\value{
A data.frame containing country names for records missing such
information.
}
\description{
Country names derived from valid geographic coordinates are added to records
missing country names.
}
\details{
This function assigns country name for records missing such
information. Country names are extracted from valid geographic coordinates
by using a high-quality map of the world (rnaturalearth package). No
country name is added to records whose coordinates are in the sea.
}
\examples{
\dontrun{
decimalLatitude <- c(-63.43333, -67.91667, -41.90000, -46.69778)
decimalLongitude <- c(-17.90000, -14.43333, -13.25000, -13.82444)
country <- c("", "NA", NA, "Brazil")

x <- data.frame(decimalLatitude, decimalLongitude, country)

bdc_country_from_coordinates(
  data = x,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  country = "country")
}

}
