% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_country_xy_mismatch.R
\name{bdc_country_xy_mismatch}
\alias{bdc_country_xy_mismatch}
\title{Identify records in a country or at a determined distance from the country
(e.g., in the ocean)}
\usage{
bdc_country_xy_mismatch(
  data,
  country_name,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  dist = 0.5
)
}
\arguments{
\item{data}{data.frame. Containing longitude and latitude}

\item{country_name}{character string. Name of the country to considered.}

\item{lon}{character string. The column with the longitude coordinates. Default = “decimallatitude”.}

\item{lat}{character string. The column with the latitude coordinates. Default = “decimallatitude”.}

\item{dist}{numeric. The distance in degrees used to created a buffer around the country.}
}
\value{
a data.frame with flagged records in column \code{.xy_out_country}
}
\description{
This function identify/flag geographic coordinates using a buffer for the
country choosed by the user. See @details.
}
\details{
Records within a buffer around a country but not in other countries
(e.g., records in the ocean) are flagged as TRUE (test passed). This avoids
to flag as FALSE records close to country limits. For example, records of
coast or marshland species.
}
\examples{
\dontrun{
decimalLatitude <- c(19.9358, -13.016667, NA, "")
decimalLongitude <- c(-40.6003, -39.6, -20.5243, NA)
x <- data.frame(decimalLatitude, decimalLongitude)
bdc_country_xy_mismatch(
  data = x,
  country_name = "Brazil",
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  dist = 0.5 # in decimal degrees
)
}
}
