% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_create_figures.R
\name{bdc_create_figures}
\alias{bdc_create_figures}
\title{Creates figures reporting the results of the bdc workflow}
\usage{
bdc_create_figures(data, workflow_step = NULL)
}
\arguments{
\item{data}{data.frame Containing the results of each test.}

\item{workflow_step}{character string. Name of the workflow step. Options available are ("prefilter", "space", and "time")}
}
\value{

}
\description{
Creates figures (i.e., barplots and maps) reporting the results of data
quality tests implemented in the bdc workflow.
}
\details{
This function creates figures based on the results of data quality
tests implemented. A pre-defined list of test names is used for creating
figures depending on the name of the workflow step informed. Figures are
saving in "Output/Figures".
}
