% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_create_report.R
\name{bdc_create_report}
\alias{bdc_create_report}
\title{Creates reports on results of data quality tests}
\usage{
bdc_create_report(data, workflow_step)
}
\arguments{
\item{data}{data frame. Containing the results of data quality tests.}

\item{workflow_step}{character string containing one of the following
options("prefiter", "taxonomy", "space" or "time")}
}
\value{

}
\description{
Creates reports on results of data quality tests
}
