% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_extract_cntr_names.R
\name{bdc_extract_cntr_names}
\alias{bdc_extract_cntr_names}
\title{Internal function. Extracts countries names in different languages from
wikipedia}
\usage{
bdc_extract_cntr_names(x)
}
\arguments{
\item{x}{}
}
\description{
Internal function. Extracts countries names in different languages from
wikipedia
}
\examples{
\dontrun{
}
}
