% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_filter_name.R
\name{bdc_filter_id}
\alias{bdc_filter_id}
\title{Internal function. Filter species information from local stored taxonomic databases}
\usage{
bdc_filter_id(id, db, version = taxadb:::latest_version())
}
\arguments{
\item{id}{character vector. Containing taxonomic code.}

\item{db}{character vector. Local taxonomic database, eg. "gbif".}

\item{version}{character vector. The year of database version, eg. "2022". The default is the latest version.}
}
\value{
This function returns a data.frame in Darwin Core format with the taxonomic information for the queried ids.
}
\description{
This function filters, by exact match, species information from taxonomic database based on a taxonomic code.
}
\details{
The function looks for an exact match of taxonomic code in the local stored database and returns the matched information
as a data.frame in Darwin Core format. Unmatched ids return NA values for all taxonomic information.
}
\examples{
\dontrun{
}
}
