% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_filter_name.R
\name{bdc_filter_name}
\alias{bdc_filter_name}
\title{Internal function. Filter species information from local stored taxonomic databases}
\usage{
bdc_filter_name(sci_name, db, db_version = taxadb:::latest_version())
}
\arguments{
\item{sci_name}{character vector. Containing scientific names.}

\item{db}{character vector. Local taxonomic database, eg. "gbif".}

\item{db_version}{character vector. The year of database version, eg. "2022".
The default is the latest version.}
}
\value{
This function returns a data.frame in Darwin Core format with the
taxonomic information for the queried species.
}
\description{
This function filters, by exact match, species information from taxonomic database.
}
\details{
The function looks for an exact match of scientific name in the local stored
database and returns the matched information
as a data.frame in Darwin Core format. Unmatched species return NA values for
all taxonomic information.
}
\examples{
\dontrun{
}
}
