% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_filter_out_flags.R
\name{bdc_filter_out_flags}
\alias{bdc_filter_out_flags}
\title{Filter out rows based on flags assigned as FALSE}
\usage{
bdc_filter_out_flags(data)
}
\arguments{
\item{data}{a data.frame with flags created by functions bdc_flag_*}
}
\description{
This functions filter out rows based on any flag column assigned as FALSE
}
\examples{
\dontrun{
data \%>\%
  bdc_flag_transposed_xy() \%>\%
  bdc_filter_out_flags()
}
}
