% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_filter_out_names.R
\name{bdc_filter_out_names}
\alias{bdc_filter_out_names}
\title{Removes records with taxonomic issues}
\usage{
bdc_filter_out_names(
  data,
  notes = c("not_found", "more_one_accepted", "no_accepted", "taxo_uncer"),
  opposite = FALSE
)
}
\arguments{
\item{data}{data frame containing a column 'notes'.}

\item{notes}{character string. A column containing notes on taxonomic standardization process. Default = "notes".}

\item{opposite}{logical. Should records with not taxonomic issue be returned.
Default = FALSE.}
}
\value{
A data frame without records with taxonomic issues. If opposite is TRUE, only records with taxonomic issues are returned.
}
\description{
This function filters out records whose scientific names contain a issue,
including names not found ('notFound'), with multiple (multipleAccepted) or
none (noAcceptedName) accepted name found.
}
\details{
The function filters out records whose notes column flags taxonomic issues such as 'noFound', 'multipleAccepted' and 'noAcceptedName'.
}
\examples{

}
