% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_flag_missing_names.R
\name{bdc_flag_missing_names}
\alias{bdc_flag_missing_names}
\title{Flag records missing scientific name}
\usage{
bdc_flag_missing_names(data, sci_name = "scientificName")
}
\arguments{
\item{data}{a data.frame containing a column of species names:}
}
\description{
This function add a new column \code{.missing_name} in the returned dataset
}
