% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_flag_missing_xy.R
\name{bdc_flag_missing_xy}
\alias{bdc_flag_missing_xy}
\title{Title Flag records missing coordinates}
\usage{
bdc_flag_missing_xy(data, lon = "decimalLongitude", lat = "decimalLatitude")
}
\arguments{
\item{data.}{data.frame containing longitude and latitude}

\item{lon.}{character string. The column with longitude coordinates}

\item{lat.}{character string. The column with latitude coordinates}
}
\value{

}
\description{
Title Flag records missing coordinates
}
