% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_flag_no_name.R
\name{bdc_flag_no_name}
\alias{bdc_flag_no_name}
\title{Flag missing scientific names}
\usage{
bdc_flag_no_name(data)
}
\arguments{
\item{data}{a data.frame with the default column names: "database_id", "scientificName", "decimalLongitude", "decimalLatitude"}
}
\description{
This function add a new column \code{.no_name} in the returned dataset
}
\examples{
\dontrun{
data \%>\%
  bdc_flag_no_name()
}
}
