% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_flag_no_xy.R
\name{bdc_flag_no_xy}
\alias{bdc_flag_no_xy}
\title{Flag missing coordinates}
\usage{
bdc_flag_no_xy(data)
}
\arguments{
\item{data}{a data.frame with the default column names: "database_id", "scientificName", "decimalLongitude", "decimalLatitude"}
}
\description{
This function add a new column \code{.no_xy} in the returned dataset
}
\examples{
\dontrun{
data \%>\%
  bdc_flag_invalid_xy()
}
}
