% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_flag_outside_xy.R
\name{bdc_flag_outside_xy}
\alias{bdc_flag_outside_xy}
\title{Flag coordinates outside the world coordinates}
\usage{
bdc_flag_outside_xy(data)
}
\arguments{
\item{data}{a data.frame with the default column names: "database_id", "scientificName", "decimalLongitude", "decimalLatitude"}
}
\description{
This function add a new column \code{.outside_xy} in the returned dataset
}
\examples{
\dontrun{
data \%>\%
  bdc_flag_outside_xy()
}
}
