% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_flag_taxo_unc.R
\name{bdc_flag_taxo_unc}
\alias{bdc_flag_taxo_unc}
\title{Title: Flag terms denoting taxonomic uncertainty (T = no terms found; F = term found)}
\usage{
bdc_flag_taxo_unc(spp_names)
}
\arguments{
\item{spp_names}{}
}
\value{

}
\description{
Title: Flag terms denoting taxonomic uncertainty (T = no terms found; F = term found)
}
