% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_flag_xy_out_country.R
\name{bdc_flag_xy_out_country}
\alias{bdc_flag_xy_out_country}
\title{Title Identify records in a country or at a determined distance from the country (e.g., in the ocean)}
\usage{
bdc_flag_xy_out_country(
  data,
  country_name,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  dist = 0.5
)
}
\arguments{
\item{data}{data.frame. Containing longitude and latitude}

\item{country_name}{character string. Name of the country to considered.}

\item{lon}{character string. The column with the longitude coordinates. Default = “decimallatitude”.}

\item{lat}{character string. The column with the latitude coordinates. Default = “decimallatitude”.}

\item{dist}{numeric. The distance in degrees used to created a buffer around the country.}
}
\value{

}
\description{
Title Identify records in a country or at a determined distance from the country (e.g., in the ocean)
}
\details{
Records within a buffer around a country but not in other countries (e.g., records in the ocean) are flagged as TRUE (test passed). This avoids to flag as FALSE records close to country limits. For example, records of coast or marshland species.
}
