% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{bdc_get_wiki_country}
\alias{bdc_get_wiki_country}
\title{Title. Wrapped funciton to get cuntry names from Wikipedia}
\usage{
bdc_get_wiki_country()
}
\value{

}
\description{
Title. Wrapped funciton to get cuntry names from Wikipedia
}
