% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{bdc_get_world_map}
\alias{bdc_get_world_map}
\title{Title: FIXEME ####################################}
\usage{
bdc_get_world_map()
}
\value{

}
\description{
Title: FIXEME ####################################
}
