% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_install_gnparser.R
\name{bdc_install_gnparser}
\alias{bdc_install_gnparser}
\title{Install gnparser software}
\usage{
bdc_install_gnparser()
}
\value{
None
}
\description{
Installs the latest version of the gnparser software, compatible with Windows, macOS, and Linux.
}
\examples{
\dontrun{
bdc_install_gnparser()
}
}
