% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_invalid_basis_of_records.R
\name{bdc_invalid_basis_of_records}
\alias{bdc_invalid_basis_of_records}
\title{Identify records from doubtful source (e.g., 'fossil', MachineObservation')}
\usage{
bdc_invalid_basis_of_records(
  data,
  basisOfRecord = "basisOfRecord",
  names_to_keep = "all"
)
}
\arguments{
\item{data}{data.frame. Containing information on basis of records.}

\item{basisOfRecord}{character string. The column name with information on
basis of records. Default = "basisOfRecord".}

\item{names_to_keep}{character string. Elements of the column BasisOfRecords
to keep. Default is 'all', which considers a selected list of recommended
standard Darwin Core classes (and their spelling variations, see details).
By default, records missing (i.e., NA) or with "unknown" information on
basis of records are kept.}
}
\value{
A data.frame contain the column '.invalid_basis_of_records'. Records
that have failed in the test are flagged as "FALSE".
}
\description{
Identify records with informed basis of records (i.e., the records type, for
example, a specimen, a human observation, or a fossil specimen) no
interpretable, that do not comply with Darwin Core recommended vocabulary, or
unrealible or unsuitable for certain analyses.
}
\details{
Users are encourage to select the set of basis of records classes
to keep. Default = c("Event","HUMAN_OBSERVATION", "HumanObservation",
"LIVING_SPECIMEN", "LivingSpecimen", "MACHINE_OBSERVATION",
"MachineObservation", "MATERIAL_SAMPLE", "O", "Occurrence",
"MaterialSample", "OBSERVATION", "Preserved Specimen",
"PRESERVED_SPECIMEN", "preservedspecimen Specimen", "Preservedspecimen",
"PreservedSpecimen", "preservedspecimen", "S", "Specimen", "Taxon",
"UNKNOWN", "", NA)
}
\examples{
\dontrun{
x <- data.frame(recod_types = c("FOSSIL_SPECIMEN", "UNKNOWN", "RON", NA, "Specimen", "PRESERVED_SPECIMEN"))
bdc_invalid_basis_of_records(data = x, basisOfRecord = "recod_types", names_to_keep = "all")
}
}
