% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_invalid_coordinates.R
\name{bdc_invalid_coordinates}
\alias{bdc_invalid_coordinates}
\title{Identify records with invalid geographic coordinates}
\usage{
bdc_invalid_coordinates(
  data,
  lat = "decimalLatitude",
  lon = "decimalLongitude"
)
}
\arguments{
\item{data}{data.frame. Containing geographical coordinates. Coordinates must be expressed in decimal degree and in WGS84.}

\item{lat}{character string. The column name with latitude. Coordinates must
be expressed in decimal degree and in WGS84. Default = "decimalLatitude".}

\item{lon}{character string. The column with longitude. Coordinates must be
expressed in decimal degree and in WGS84. Default = "decimalLongitude".}
}
\value{
A data.frame contain the column '.invalid_coordinates'. Records that
have failed in the test are flagged as "FALSE".
}
\description{
Flags records with out-of-range coordinates (-90 to 90 for latitude -180 to 180 for longitude).
}
\examples{
\dontrun{
decimalLatitude <- c(-185.111, -43.34, "", -21.8069444)
decimalLongitude <- c(-45.4, -39.6, -20.5243, -440.9055555)
x <- data.frame(decimalLatitude, decimalLongitude)

bdc_invalid_coordinates(
data = x, 
lat = "decimalLatitude", 
lon = "decimalLongitude")
}
}
