% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_invalid_xy.R
\name{bdc_invalid_xy}
\alias{bdc_invalid_xy}
\title{Identify records with invalid geographic coordinates}
\usage{
bdc_invalid_xy(data, lat = "decimalLatitude", lon = "decimalLongitude")
}
\arguments{
\item{data}{data.frame. Containing geographical coordinates.}

\item{lat}{character string. The column with latitude Default =
"decimalLatitude".}

\item{lon}{character string. The column with longitude. Default =
"decimalLongitude".}
}
\value{
A data.frame contain the column '.invalid_xy'. Records that have
failed in the test are flagged as "FALSE".
}
\description{
Flags records with out-of-range coordinates (-90 to 90 for latitude -180 to 180 for longitude).
}
\examples{
\dontrun{
decimalLatitude <- c(-185.111, -43.34, "", -21.8069444)
decimalLongitude <- c(-45.4, -39.6, -20.5243, -440.9055555)
x <- data.frame(decimalLatitude, decimalLongitude)
bdc_invalid_xy(data = x, lat = "decimalLatitude", lon = "decimalLongitude")
}
}
