% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_missing_coordinates.R
\name{bdc_missing_coordinates}
\alias{bdc_missing_coordinates}
\title{Identify records missing geographic coordinates}
\usage{
bdc_missing_coordinates(
  data,
  lat = "decimalLatitude",
  lon = "decimalLongitude"
)
}
\arguments{
\item{data}{data.frame. Containing geographical coordinates.}

\item{lat}{character string. The column name with latitude. Coordinates must
be expressed in decimal degree and in WGS84. Default = "decimalLatitude".}

\item{lon}{character string. The column with longitude. Coordinates must be
expressed in decimal degree and in WGS84. Default = "decimalLongitude".}
}
\value{
A data.frame contain the column '.missing_coordinates'. Records that
have failed in the test are flagged as "FALSE".
}
\description{
This function flags records missing latitude or longitude coordinates.
}
\details{
This test identifies records missing geographic coordinates (i.e.,
empty or not applicable \link{NA} longitude or latitude)
}
\examples{
\dontrun{
decimalLatitude <- c(19.9358, -13.016667, NA, "")
decimalLongitude <- c(-40.6003, -39.6, -20.5243, NA)
x <- data.frame(decimalLatitude, decimalLongitude)

bdc_missing_coordinates(
data = x, 
lat = "decimalLatitude", 
lon = "decimalLongitude")
}
}
