% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_missing_names.R
\name{bdc_missing_names}
\alias{bdc_missing_names}
\title{Identify records missing scientific name}
\usage{
bdc_missing_names(data, sci_names = "scientificName")
}
\arguments{
\item{data}{data.frame. Containing species scientific names.}

\item{sci_names}{character string. The column name with species scientific
name. Default = "scientificName".}
}
\value{
A data.frame contain the column '.missing_name'. Records that have
failed in the test are flagged as "FALSE".
}
\description{
Flags records with missing or not interpretable scientific names.
}
\details{
This test identifies records missing scientific names (i.e., empty
or not applicable \link{NA} names)
}
\examples{
\dontrun{
x <- data.frame(scientificName = c("Ocotea odorifera", NA, "Panthera onca", ""))
bdc_missing_names(data = x, sci_names = "scientificName")
}
}
