% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_missing_xy.R
\name{bdc_missing_xy}
\alias{bdc_missing_xy}
\title{Identify records missing geographic coordinates}
\usage{
bdc_missing_xy(data, lon = "decimalLongitude", lat = "decimalLatitude")
}
\arguments{
\item{data}{data.frame. Containing geographical coordinates.}

\item{lon}{character string. The column with longitude. Default =
"decimalLongitude".}

\item{lat}{character string. The column with latitude Default =
"decimalLatitude".}
}
\value{
A data.frame contain the column '.missing_xy'. Records that have
failed in the test are flagged as "FALSE".
}
\description{
Flags records missing latitude or longitude coordinates.
}
\details{
This test identifies records missing geographic coordinates (i.e.,
empty or not applicable \link{NA} longitude or latitude)
}
\examples{
\dontrun{
decimalLatitude <- c(19.9358, -13.016667, NA, "")
decimalLongitude <- c(-40.6003, -39.6, -20.5243, NA)
x <- data.frame(decimalLatitude, decimalLongitude)
bdc_missing_xy(data = x, lat = "decimalLatitude", lon = "decimalLongitude")
}
}
