% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_parse_date.R
\name{bdc_parse_date}
\alias{bdc_parse_date}
\title{Extract and flag year from date}
\usage{
bdc_parse_date(data, col_to_test, year_threshold = NULL)
}
\arguments{
\item{x:}{A data frame containing column of event date.}

\item{col_to_test:}{Numeric or date. The column with event date information.}

\item{year_threshold:}{Numeric. A years threshold (four digit year) to flag records due to their age. Default = NULL.}
}
\value{
A data frame with the original data (x), a flagged column (.year) and a column with the extracted four digit years (year). Records with .year = FALSE means dates older than year threshold or 1500.
}
\description{
Extract and flag year from date
}
\details{
The function filters the column containing event dates and extract the dates with four digits. If the event dates are older than 1500, or the threshold year, they are flagged FALSE.
}
