% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{bdc_query_wfo}
\alias{bdc_query_wfo}
\title{Title: Query names in WorldFlora Online database}
\usage{
bdc_query_wfo(species_name, match_dist)
}
\arguments{
\item{species_name}{}

\item{match_dist}{}
}
\value{

}
\description{
Title: Query names in WorldFlora Online database
}
