% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_quickmap.R
\name{bdc_quickmap}
\alias{bdc_quickmap}
\title{Create a map of points using ggplot2}
\usage{
bdc_quickmap(data, lat, lon, col_to_map = NULL, size = size)
}
\arguments{
\item{data}{data.frame. Containing longitude and latitude}

\item{lat}{character string. The column with the latitude coordinates}

\item{lon}{character string. The column with the longitude coordinates}

\item{col_to_map}{character string. Defining the column or color used to map.
Can be a color name (e.g "red") the the name of a column of data. Default =
"blue"}

\item{size}{numeric. The size of the points.}
}
\description{
Creates a map of points using ggplot2 useful for inspecting the results of
tests implemented in the bdc workflow.
}
\details{
Only records with valid coordinates can be plotted. Records missing
or with invalid coordinates are removed prior creating the map.
}
\examples{
\dontrun{
decimalLatitude <- c(19.9358, -13.016667, -19.935800)
decimalLongitude <- c(-40.6003, -39.6, -40.60030)
.coordinates_out_country <- c(FALSE, TRUE, TRUE)
x <- data.frame(decimalLatitude, decimalLongitude, .coordinates_out_country)

bdc_quickmap(
  data = x,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  col_to_map = ".coordinates_out_country",
  size = 1)
}
}
