% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_rem_family_names.R
\name{bdc_rem_family_names}
\alias{bdc_rem_family_names}
\title{Title: Remove family names from species names}
\usage{
bdc_rem_family_names(sp_names)
}
\arguments{
\item{sp_names}{}
}
\value{

}
\description{
Title: Remove family names from species names
}
