% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{bdc_rem_other_issues}
\alias{bdc_rem_other_issues}
\title{Title: Remove duplicated genus, substitute empty cells by NA, capitalize generic name}
\usage{
bdc_rem_other_issues(spp_names)
}
\arguments{
\item{spp_names}{}
}
\value{

}
\description{
Title: Remove duplicated genus, substitute empty cells by NA, capitalize generic name
}
