% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{bdc_rem_taxo_unc}
\alias{bdc_rem_taxo_unc}
\title{Title: Remove terms denoting taxonomic uncertainty (e.g., cf., sp., aff)}
\usage{
bdc_rem_taxo_unc(spp_names)
}
\arguments{
\item{spp_names}{}
}
\value{

}
\description{
Title: Remove terms denoting taxonomic uncertainty (e.g., cf., sp., aff)
}
