% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_scientificName_empty.R
\name{bdc_scientificName_empty}
\alias{bdc_scientificName_empty}
\title{Identify records with empty scientific names}
\usage{
bdc_scientificName_empty(data, sci_names = "scientificName")
}
\arguments{
\item{data}{data.frame. Containing species scientific names.}

\item{sci_names}{character string. The column name with species scientific
name. Default = "scientificName".}
}
\value{
A data.frame contain the column ".scientificName_empty". Compliant
(TRUE) if 'sci_names' is not empty; otherwise "FALSE".
}
\description{
Flags records with empty or not interpretable scientific names.
}
\details{
This test identifies records missing scientific names (i.e., empty
or not applicable \link{NA} names)
}
\examples{
\dontrun{
x <- data.frame(scientificName = c("Ocotea odorifera", NA, "Panthera onca", ""))
bdc_scientificName_empty(data = x, sci_names = "scientificName")
}
}
