% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{bdc_standard_country}
\alias{bdc_standard_country}
\title{Title: standard_country is a function to correct, standardize, and assign a ISO code to country names}
\usage{
bdc_standard_country(data, cntry, cntry_names_db)
}
\arguments{
\item{data}{}

\item{cntry}{}

\item{cntry_names_db}{}
}
\value{

}
\description{
Title: standard_country is a function to correct, standardize, and assign a ISO code to country names
}
