% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_standardize_country.R
\name{bdc_standardize_country}
\alias{bdc_standardize_country}
\title{Title: standard_country is a function to correct, standardize, and assign a ISO code to country names.}
\usage{
bdc_standardize_country(data, country, country_names_db)
}
\arguments{
\item{data}{A data.frame with a column of country names.}

\item{country}{Column name with the country names.}

\item{country_names_db}{A database to match country names and return the suggested one.}
}
\value{
Return a data.frame with original country names, suggested names and ISO code to country names.
}
\description{
Title: standard_country is a function to correct, standardize, and assign a ISO code to country names.
}
\examples{
\dontrun{
data \%>\%
  bdc_flag_transposed_xy() \%>\%
  bdc_filter_out_flags()
}
}
