% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{bdc_standardize_dataset}
\alias{bdc_standardize_dataset}
\title{Title: Standardize datasets columns based on metadata file}
\usage{
bdc_standardize_dataset(metadata)
}
\arguments{
\item{metadata}{a table containing information about which columns of the
original dataset need to be renamed following Darwin Core terminology.
Please see the \code{Config/DatabaseInfo.csv} file.}
}
\description{
Title: Standardize datasets columns based on metadata file
}
