% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_summary_col.R
\name{bdc_summary_col}
\alias{bdc_summary_col}
\title{Creates or updates the column summarizing the results of data quality tests}
\usage{
bdc_summary_col(data)
}
\arguments{
\item{data}{data.frame. Containing the results of data quality tests (i.e.,
columns starting with ".").}
}
\value{
A clean database without records flagged as "FALSE".
}
\description{
Creates or updates a column summarizing the results of data quality tests
(i.e., columns starting with "."). Records that have failed in at least one
data quality test are considered invalid (i.e., flagged as "FALSE") in the
".summary" column.
}
\details{
If existing, the column ".summary" will be removed and then updated
considering all test names available in the supplied database.
}
