% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_transposed_coordinates.R
\name{bdc_transposed_coordinates}
\alias{bdc_transposed_coordinates}
\title{Identify transposed geographic coordinates}
\usage{
bdc_transposed_coordinates(
  data,
  id = "database_id",
  sci_names = "scientificName",
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  country = "country",
  countryCode = "countryCode"
)
}
\arguments{
\item{data}{data.frame. Containing a unique identifier for each records,
geographical coordinates, and country names. Coordinates must be expressed in
decimal degree and in WGS84.}

\item{id}{character string. The column name with a unique record identifier.
Default = "database_id".}

\item{sci_names}{character string. The column name with species scientific
name. Default = "scientificName".}

\item{lat}{character string. The column name with latitude. Coordinates must
be expressed in decimal degree and in WGS84. Default = "decimalLatitude".}

\item{lon}{character string. The column with longitude. Coordinates must be
expressed in decimal degree and in WGS84. Default = "decimalLongitude".}

\item{country}{character string. The column name with the country
assignment of each record. Default = "country".}
}
\value{
A data.frame containing the column 'transposed_coordinates'. Records
that have failed in the test are flagged as "FALSE".
}
\description{
This function flags and corrects records when latitude and longitude appear
to be transposed.
}
\details{
This test identifies transposed coordinates resulted from mismatches
between the country informed to a record and coordinates. Transposed
coordinates often fall outside of the indicated country (i.e., in other
countries or in the sea). Different coordinate transformations are
performed to correct country/coordinates mismatch. Importantly, verbatim
coordinates are replaced by the corrected ones in the returned database. A
database containing verbatim and corrected coordinates is created in
"Output/Check/01_transposed_coordinates.csv".
}
\examples{
\dontrun{
id <- c(1,2,3,4)
scientificName <- c("Rhinella major", "Scinax ruber", 
                    "Siparuna guianensis", "Psychotria vellosiana")
decimalLatitude <- c(-63.43333, -67.91667, -41.90000, -46.69778)
decimalLongitude <- c(-17.90000, -14.43333, -13.25000, -13.82444)
country <- c("BOLIVIA", "bolivia", "Brasil", "Brazil")

x <- data.frame(id, scientificName, decimalLatitude,
                decimalLongitude, country)

bdc_transposed_coordinates(
  data = x,
  id = "id",
  sci_names = "scientificName",
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  country = "country")
}

}
