% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_transposed_xy.R
\name{bdc_transposed_xy}
\alias{bdc_transposed_xy}
\title{Identify transposed geographic coordinates and standardizes country names}
\usage{
bdc_transposed_xy(
  data,
  id,
  sci_names,
  lat,
  lon,
  country,
  country_from_xy = TRUE
)
}
\arguments{
\item{data}{data.frame. Containing an unique identifier for each records,
geographical coordinates, and country names.}

\item{id}{character string. The column with an unique record identifier.
Default = "database_id".}

\item{sci_names}{character string. The column with species scientific name.
Default = "scientificName".}

\item{lat}{character string. The column with latitude Default =
"decimalLatitude".}

\item{lon}{character string. The column with longitude. Default =
"decimalLongitude".}

\item{country}{character string. The column with the country assignment of
each record. Default = "country".}
}
\value{
A data.frame containing three columns i) 'transposed_xy' (logical,
records that have failed in the test are flagged as "FALSE", ii)
country_suggested (standardized country names), and iii) country_code
(standardized iso3 country code).
}
\description{
Flags and corrects records when latitude and longitude appear to be
transposed and standardizes country names.
}
\details{
This test identifies transposed coordinates resulted from mismatches
between the country informed to a record and coordinates. Transposed
coordinates often fall outside of the indicated country (i.e., in other
countries or in the sea). If 'country_from_xy' is TRUE country names
derived from valid coordinates is added to records missing country
information. Otherwise, only records with interpretable country names are
considered. First, country names are standardized using an exact matching
against a list of country names in several languages from Wikipedia. If
there any unmatched names remain, a fuzzy matching algorithm is used to
find potential candidates for each misspelled countries names.  Then,
standardized country names are used to identifying mismatches between
country and coordinates. To do this, the 'seas' and, 'countries' tests from
'CoordinateCleaner' package are used. Once detected a mismatch, different
coordinate transformations are performed to correct country/coordinates
mismatch. Importantly, verbatim coordinates are replaced by the corrected
ones in the returned database. A database containing verbatim and corrected
coordinates is created in "Output/Check/01_transposed_xy.csv".
}
\examples{
id <- c(1,2,3,4)
scientificName <- c("Rhinella major", "Scinax ruber", 
                    "Siparuna guianensis", "Psychotria vellosiana")
decimalLatitude <- c(-63.43333, -67.91667, -41.90000, -46.69778)
decimalLongitude <- c(-17.90000, -14.43333, -13.25000, -13.82444)
country <- c("BOLIVIA", "bolivia", "Brasil", "Brazil")

x <- data.frame(id, scientificName, decimalLatitude,
                decimalLongitude, country)

bdc_transposed_xy(
  data = x,
  id = "id",
  sci_names = "scientificName",
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  country = "country"
)
}
