% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_xy_from_locality.R
\name{bdc_xy_from_locality}
\alias{bdc_xy_from_locality}
\title{Identify records missing coordinates but with information on locality}
\usage{
bdc_xy_from_locality(
  data,
  locality = "locality",
  lon = "decimalLongitude",
  lat = "decimalLatitude"
)
}
\arguments{
\item{data}{data.frame. Containing geographical coordinates and the column
"locality'.}

\item{locality}{chracter string. The column with information on where records
was collected.}

\item{lon}{character string. The column with longitude. Default =
"decimalLongitude".}

\item{lat}{character string. The column with latitude Default =
"decimalLatitude".}
}
\value{

}
\description{
Identify records whose coordinates can potentially be extracted from
information on locality
}
