% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_xy_precision.R
\name{bdc_xy_precision}
\alias{bdc_xy_precision}
\title{Flags imprecise geographic coordinates}
\usage{
bdc_xy_precision(
  data,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  ndec = c(0, 1, 2)
)
}
\arguments{
\item{data:}{data.frame. A data.frame containing geographic coordinates.}

\item{lon:}{character string. The column with longitude. Default =
"decimalLongitude".}

\item{lat:}{character string. The column with latitude Default =
"decimalLatitude".}

\item{ndec:}{numeric. Containing the numbers of decimals place to be tested.
Default = c(0,1,2).}
}
\value{
A data.frame with logical values indicating whether values are
rounded by the specified decimal number (ndec). In other word, potentially
imprecise coordinates.
}
\description{
This function is used to flag records with a coordinate precision below a
specified number of decimal places. Coordinates with one, two, or three
decimal places have precision of~11.1 km, ~1.1 km, and ~111 m at the equator,
respectively.
}
\examples{
"data <- data.frame(lon = c(-55.389, -13.897, 30.678, 90.675) , lat = c(-21.345, 23.567, 16.798, -10.468))"
"bdc_round_dec(data = data, lon = "lon", lat = "lat", ndec = c(0, 2))"

}
