% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{biotime_deu}
\alias{biotime_deu}
\title{bioTime data for Germany}
\format{A \code{data.frame} with 58739 observation and 15 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://zenodo.org/record/3265871}}
}
\usage{
data(biotime_deu)
}
\description{
biodiversity time series trends from the BioTime database for Germany
}
\details{
Biodiversity time series trends extracted from the BioTime database for the 
extent of Germany. For Bavaria no data is available.
}
\references{
Dornelas M, Antão LH, Moyes F, Bates, AE, Magurran, AE, et al. BioTIME: 
A database of biodiversity time series for the Anthropocene. 
Global Ecol Biogeogr. 2018; 27:760 - 786. https://doi.org/10.1111/geb.12729
}
