% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{bird_bav}
\alias{bird_bav}
\title{bird range polygon data for Bavaria}
\format{A \code{data.frame} with 283 observation and 18 variables.}
\source{
This data has been obtained from: \itemize{\item \url{http://datazone.birdlife.org/species/requestdis}}
}
\usage{
data(bird_bav)
}
\description{
range polygons of all bird species for Bavaria
}
\details{
Bird range polygons of all bird species for Bavaria obtained from BirdLife.
}
