% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{bird_bva_shape}
\alias{bird_bva_shape}
\title{breeding bird ranges of Bavaria}
\format{A \code{data.frame} with 57856 observation and 38 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://www.lfu.bayern.de/natur/atlas_brutvoegel/index.htm}}
}
\usage{
data(bird_bva_shape)
}
\description{
range points of saP-relevant breeding bird species for Bavaria
}
\details{
Range points of Bavarian saP-relevant breeding bird species
}
