% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{bird_bva_shape_tk4tel}
\alias{bird_bva_shape_tk4tel}
\title{breeding bird ranges of Bavaria}
\format{
A \code{data.frame} with 1980 observation and 150 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://www.lfu.bayern.de/natur/atlas_brutvoegel/index.htm}}
}
\usage{
data(bird_bva_shape_tk4tel)
}
\description{
range points of saP-relevant breeding bird species for Bavaria
}
\details{
Bavarian saP-relevant breeding bird species data gridded to TK4tel
}
