% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{carbon_bav}
\alias{carbon_bav}
\title{Carbon stock data for Bavaria}
\format{A \code{data.frame} with 1247 observation and 7 variables.}
\usage{
data(carbon_bav)
}
\description{
Potential and actual carbon stock for Bavaria
}
\details{
Potential and actual carbon stock for Bavaria at 5 Arc min resolution.
}
\references{
Erb, Karl-Heinz, Tamara Fetzel, Christoph Plutzar, 
Thomas Kastner, Christian Lauk, Andreas Mayer, 
Maria Niedertscheider, Christian Korner, and Helmut Haberl. 
Biomass Turnover Time in Terrestrial Ecosystems Halved by Land Use. 
Nature Geoscience 9, no. 9 (September 2016): 674-78.
}
