% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{cci_bav_tk25}
\alias{cci_bav_tk25}
\title{CCI land cover data of bavaria}
\format{A \code{data.frame} with 2268 observations and 26 variables.}
\usage{
data(cci_bav_tk25)
}
\description{
CCI land cover data of Bavaria, Germany
}
\details{
CCI land cover data of Bavaria, Germany resampled onto the TK25 grid. 
CCI Land cover are global land cover maps at 300m spatial resolution based on MERIS and SPOT satellite data. 
Data includes annual information from 1992 - 2015.
}
