% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{chelsa_bav_tk4tel}
\alias{chelsa_bav_tk4tel}
\title{Chelsa climate data of bavaria}
\format{
A \code{data.frame} with 2268 observations and 66 variables.
}
\source{
This data has been obtained from: \url{http://chelsa-climate.org/}.
}
\usage{
data(chelsa_bav_tk4tel)
}
\description{
Chelsa climate data of Bavaria, Germany
}
\details{
Chelsa climate data of Bavaria, Germany resampled onto the TK4tel grid. 
The data.frame contains bioclimatic variables, monthly precipitation,
monthly minimum, mean and maximum temperature of current climatologies (1979-2013).
}
\references{
Karger et al. (2017) Climatologies at high resolution for 
the earth's land surface areas. Scientific Data 4, 170122.
}
