% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_clean_duplicates.R
\name{clean_duplicates}
\alias{clean_duplicates}
\title{Title: Filter duplicated names in bdc_get_taxa_taxadb function.
Function only to internal use of bdc_get_taxa_taxadb function.}
\usage{
clean_duplicates(data)
}
\arguments{
\item{data}{A data.frame created in bdc_get_taxa_taxadb.}
}
\value{
Return the data.frame without duplicates.
}
\description{
Title: Filter duplicated names in bdc_get_taxa_taxadb function.
Function only to internal use of bdc_get_taxa_taxadb function.
}
