% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{cordex_tasminAdjust_bav}
\alias{cordex_tasminAdjust_bav}
\title{Euro-cordex minimum temperature data of bavaria}
\format{A \code{data.frame} with 37837287 observations and 9 variables.}
\usage{
data(cordex_tasminAdjust_bav)
Euro-cordex climate simulation data of Bavaria, Germany. 
The data.frame contains monthly bias-adjusted minimum temperature data from 1960 - 2099 under 3 representative concentration 
pathways (rcps; RCP2.6, RCP4.5, RCP8.5), 5 global circulation models (gcms; MPI-M-MPI-ESM-LR, CNRM-CERFACS-CNRM-CM5, 
MOHC-HadGEM2-ES, ICHEC-EC-EARTH, IPSL-IPSL-CM5A-MR), 5 regional climate models (rcms; CLMcom-CCLM4-8-17, 
KNMI-RACMO22E, MPI-CSC-REMO2009, SMHI-RCA4, DMI-HIRHAM5), 4 ensemble (r1i1p1, r2i1p1, r3i1p1, r12i1p1) and 3 rs (v1, v1a, v2).
}
\description{
Euro-cordex monthly minimum temperature data of Bavaria, Germany
}
