% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{corine_cha_bav}
\alias{corine_cha_bav}
\title{Corine land cover change data of bavaria}
\format{
A \code{data.frame} with 149217 observations and 10 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://land.copernicus.eu/pan-european/corine-land-cover}}
}
\usage{
data(corine_cha_bav)
}
\description{
Corine land cover change data of Bavaria, Germany
}
\details{
Corine land cover change data of Bavaria, Germany. 
Corine Land cover change are land cover change maps for Europe mapped at a resolution of 100 x 100 m.
}
