% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{corine_cha_bav_tk25}
\alias{corine_cha_bav_tk25}
\title{Chelsa climate data of bavaria}
\format{A \code{data.frame} with 2056 observations and 6 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://land.copernicus.eu/pan-european/corine-land-cover}}
}
\usage{
data(corine_cha_bav_tk25)
}
\description{
Chelsa climate data of Bavaria, Germany
}
\details{
Corine land cover change data of Bavaria, Germany resampled onto the TK25 grid. 
Corine Land cover change are land cover change maps for Europe mapped at a resolution of 5 ha derived 
from the raw Corine land cover data.
}
