% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{corine_cha_bav_tk4tel}
\alias{corine_cha_bav_tk4tel}
\title{Corine land cover change data of bavaria}
\format{
A \code{data.frame} with 2268 observations and 7 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://land.copernicus.eu/pan-european/corine-land-cover}}
}
\usage{
data(corine_lc_bav_tk4tel)
}
\description{
Corine land cover change data of Bavaria, Germany
}
\details{
Corine land cover change data of Bavaria, Germany resampled onto the TK4tel grid. 
Corine Land cover change are land cover change maps for Europe for 1990, 2000, 2006, 2012, and 2018. 
It consists of an inventory of land cover in 44 classes.
}
