% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{corine_cha_perc_bav}
\alias{corine_cha_perc_bav}
\title{Percentage cover of corine land cover change classes of bavaria}
\format{
A \code{data.frame} with 30864 observations and 33 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://land.copernicus.eu/pan-european/corine-land-cover}}
}
\usage{
data(corine_cha_perc_bav)
}
\description{
Percentage cover of Corine land cover change classes of Bavaria, Germany
}
\details{
Percentage cover of Corine land cover change classes of Bavaria, Germany.
Percentage cover of Corine Land cover change was derived from raw data and aggregated to a resolution of 1 x 1 km.
}
