% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{corine_lc_bav}
\alias{corine_lc_bav}
\title{Corine land cover data of bavaria}
\format{
A \code{data.frame} with 7054468 observations and 7 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://land.copernicus.eu/pan-european/corine-land-cover}}
}
\usage{
data(corine_lc_bav)
}
\description{
Corine land cover data of Bavaria, Germany
}
\details{
Corine land cover data of Bavaria, Germany. 
Corine Land cover are land cover maps for Europe mapped at a resolution of 100 x 100 m.
}
