% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{dist_pa_bav_tk4tel}
\alias{dist_pa_bav_tk4tel}
\title{Distance to protected area data of bavaria}
\format{
A \code{data.frame} with 22979 observations and 4 variables.
}
\source{
Cite this dataset as: \itemize{\item UNEP-WCMC and IUCN (2021), Protected Planet: 
The World Database on Protected Areas (WDPA) On-line, June/2021, Cambridge, UK: UNEP-WCMC and IUCN. 
Available at: \url{https://protectedplanet.net.}.}
}
\usage{
data(dist_pa_bav_tk4tel)
}
\description{
Distance to IUCN protected area data of Bavaria, Germany
}
\details{
Distance to protected areas of Bavaria, Germany derived from the World Database on 
Protected Areas and resampled onto the TK4tel grid.
}
