% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{diva_cover_deu}
\alias{diva_cover_deu}
\title{Diva-GIS land cover data of germany}
\format{A \code{data.frame} with 1071360 observations and 3 variables.}
\source{
This data has been obtained from: \itemize{\item \url{http://www.diva-gis.org/gdata}}
}
\usage{
data(diva_cover_deu)
}
\description{
Diva-GIS land cover data of Germany
}
\details{
Land cover data of Germany downloaded from diva-gis.org. 
Land cover comes originally from GLC2000 and has been resampled onto a 30 seconds grid.
}
