% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{diva_pop_deu}
\alias{diva_pop_deu}
\title{Diva-GIS population data of germany}
\format{A \code{data.frame} with 189530 observations and 3 variables.}
\source{
This data has been obtained from: \itemize{\item \url{http://www.diva-gis.org/gdata}}
}
\usage{
data(diva_pop_deu)
}
\description{
Diva-GIS population data of Germany
}
\details{
Population data (old) of Germany downloaded from diva-gis.org, which is originally derived from 
the global gridded population database (CIESIN, 2000) and has a resoltuon of 30 Arc seconds.
}
