% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{diva_rails_deu}
\alias{diva_rails_deu}
\title{Diva-GIS railroads data of germany}
\format{A \code{sf} object with 5268 observations and 8 variables.}
\source{
This data has been obtained from: \itemize{\item \url{http://www.diva-gis.org/gdata}}
}
\usage{
data(diva_rails_deu)
}
\description{
Diva-GIS railroads data of Germany
}
\details{
Railroads data of Germany downloaded from diva-gis.org derived from
the Digital Chart of the World.
}
