% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{diva_roads_deu}
\alias{diva_roads_deu}
\title{Diva-GIS roads data of germany}
\format{A \code{sf} object with 4713 observations and 6 variables.}
\source{
This data has been obtained from: \itemize{\item \url{http://www.diva-gis.org/gdata}}
}
\usage{
data(diva_roads_deu)
}
\description{
Diva-GIS roads data of Germany
}
\details{
Roads data of Germany downloaded from diva-gis.org derived from
the Digital Chart of the World.
}
