% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{diva_water_areas_deu}
\alias{diva_water_areas_deu}
\title{Diva-GIS inland water data of germany}
\format{A \code{sf} object with 907 observations and 6 variables.}
\source{
This data has been obtained from: \itemize{\item \url{http://www.diva-gis.org/gdata}}
}
\usage{
data(diva_water_areas_deu)
}
\description{
Diva-GIS inland water data of Germany (area features)
}
\details{
Rivers, canals, and lakes of Germany downloaded from diva-gis.org derived from
the Digital Chart of the World. Seperate files for line and area features
}
