% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{eurolst_bav_tk25}
\alias{eurolst_bav_tk25}
\title{EuroLST climate data of bavaria}
\format{A \code{data.frame} with 2268 observations and 11 variables.}
\source{
This data has been obtained from: \itemize{\item \url{http://www.geodati.fmach.it/eurolst.html}}
}
\usage{
data(eurolst_bav_tk25)
}
\description{
EuroLST climate data of Bavaria, Germany
}
\details{
Bioclimatic data (bio1 - bio11) derived from MODIS Satellite images
resampled onto the TK25 grid.
}
\references{
Metz, M.; Rocchini, D.; Neteler, M. 2014: Surface temperatures at the continental scale: 
Tracking changes with remote sensing at unprecedented detail. Remote Sensing. 2014, 6(5): 3822-3840
}
