% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{ewembi_bav_tk25}
\alias{ewembi_bav_tk25}
\title{EWEMBI climate data of bavaria}
\format{A \code{data.frame} with 9072 observations and 13883 variables.}
\usage{
data(ewembi_bav_tk25)
}
\description{
EWEMBI climate data of Bavaria, Germany
}
\details{
Daily precipitation and minimum, maximum and mean temperature of Bavaria, Germany from 1979 - 2016
at a spatial resolution of 0.5 degree resampled onto the TK25 grid.
}
