% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{forest_disturbance_bav_300m}
\alias{forest_disturbance_bav_300m}
\title{Forest disturbance data of bavaria}
\format{
A \code{data.frame} with 783599 observations and 5 variables.
}
\usage{
data(forest_disturbance_bav_300m)
}
\description{
Forest disturbance data of Bavaria, Germany summarised to 300 m resolution
}
\details{
Forest disturbance data of Bavaria, Germany derived from Landsat Imager 
at a spatial resolution of 30 m resampled to 300 m resolution.
}
