% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{gard_reptiles_bav}
\alias{gard_reptiles_bav}
\title{Reptile range data of bavaria}
\format{A \code{data.frame} with 12 observations and 7 variables.}
\usage{
data(gard_reptiles_bav)
}
\description{
Gard reptile range data of Bavaria, Germany
}
\details{
GARD reptile range data of Bavaria, Germany
}
