% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{gimms3g_v0_bav}
\alias{gimms3g_v0_bav}
\title{GIMMS3g v0 ndvi data of bavaria}
\format{A \code{data.frame} with 973340 observations and 5 variables.}
\usage{
data(gimms3g_v0_bav)
}
\description{
GIMMS3g v0 ndvi data of Bavaria, Germany
}
\details{
GIMMS3g v0 ndvi data of of Bavaria, Germany from 1981 to 2013.
}
