% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{gimms3g_v1_bav}
\alias{gimms3g_v1_bav}
\title{GIMMS3g v1 ndvi data of bavaria}
\format{A \code{data.frame} with 1033344 observations and 4 variables.}
\usage{
data(gimms3g_v1_bav)
}
\description{
GIMMS3g v1 ndvi data of Bavaria, Germany
}
\details{
GIMMS3g v0 ndvi data of of Bavaria, Germany from 1981 to 2015.
}
