% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{globcover_bav_tk4tel}
\alias{globcover_bav_tk4tel}
\title{Globcover land cover data of bavaria}
\format{
A \code{data.frame} with 2268 observations and 6 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{http://due.esrin.esa.int/page_globcover.php}}
}
\usage{
data(globcover_bav_tk4tel)
}
\description{
Globcover land cover data of Bavaria, Germany
}
\details{
Land cover data of Bavaria, Germany for 2004 and 2009
resampled onto the TK4tel grid.
}
