% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{hfp_1993_v3_bav}
\alias{hfp_1993_v3_bav}
\title{Human footprint of bavaria in 1993}
\format{A \code{data.frame} with 70330 observations and 11 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://wcshumanfootprint.org/}}
}
\usage{
data(hfp_1993_v3_bav)
}
\description{
Global human footprint v3 for 1993 of Bavaria, Germany
}
\details{
Global human footprint v3 for 1993 of Bavaria, Germany
}
