% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{iba_bav}
\alias{iba_bav}
\title{Important Bird Areas of bavaria}
\format{A \code{sf} object with 157 observations and 10 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://bergenhusen.nabu.de/forschung/ibas/index.html}}
}
\usage{
data(iba_bav)
}
\description{
Important Bird Areas of Bavaria, Germany
}
\details{
Important Bird Area vector data of Bavaria, Germany.
}
