% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{isimip_lu_bav_tk25}
\alias{isimip_lu_bav_tk25}
\title{ISIMIP2b land-use data of bavaria}
\format{A \code{data.frame} with 36288 observations and 31 variables.}
\usage{
data(isimip_lu_bav_tk25)
}
\description{
ISIMIP2b land-use data of Bavaria, Germany for 1995 and 2080
}
\details{
Land-use data of Bavaria, Germany for 1995 and 2080 
under two representative concentration pathways (RCP2.6 and RCP8.0) and 
4 global climate models (GFDL-ESM2M, HadGEM2-ES, IPSL-CM5A-LR, MIROC5) and 
resampled onto the TK25 grid.
}
