% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{kk09_bav}
\alias{kk09_bav}
\title{KK09 data of Bavaria}
\format{A \code{data.frame} with 3741 observations and 289 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://doi.org/10.1594/PANGAEA.893758}}
}
\usage{
data(kk09_bav)
}
\description{
KK09 data of Bavaria, Germany
}
\details{
The KK09 Anthropogenic Land Cover Change Scenarios for Bavaria, Germany
}
\references{
Kaplan, Jed O; Krumhardt, Kristen M; Zimmermann, Niklaus E (2009): 
The prehistoric and preindustrial deforestation of Europe. 
Quaternary Science Reviews, 28(27-28), 3016-3034, https://doi.org/10.1016/j.quascirev.2009.09.028
}
