% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{kk10_bav}
\alias{kk10_bav}
\title{KK10 data of Bavaria}
\format{A \code{data.frame} with 1247 observations and 7903 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://doi.org/10.1594/PANGAEA.871369}}
}
\usage{
data(kk10_bav)
}
\description{
KK10 data of Bavaria, Germany
}
\details{
The KK10 Anthropogenic Land Cover Change scenario for the preindustrial Holocene data of Bavaria, Germany
}
\references{
Kaplan, Jed O; Krumhardt, Kristen M; Ellis, Erle C; Ruddiman, William F; 
Lemmen, Carsten; Klein Goldewijk, Kees (2011): Holocene carbon emissions as a result of 
anthropogenic land cover change. The Holocene, 21(5), 775-791, https://doi.org/10.1177/0959683610386983
}
