% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{lakes_points_bav}
\alias{lakes_points_bav}
\title{Lake point data of Bavaria}
\format{A \code{data.frame} with 423 observations and 22 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://www.hydrosheds.org/downloads}}
}
\usage{
data(lakes_points_bav)
}
\description{
Lake point data of Bavaria, Germany
}
\details{
Lake point data of Bavaria, Germany
}
