% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{lakes_poly_bav}
\alias{lakes_poly_bav}
\title{Lake polygon data of Bavaria}
\format{A \code{data.frame} with 427 observations and 22 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://www.hydrosheds.org/downloads}}
}
\usage{
data(lakes_poly_bav)
}
\description{
Lake polygon data of Bavaria, Germany
}
\details{
Lake polygon data of Bavaria, Germany
}
