% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{landsystem_bav}
\alias{landsystem_bav}
\title{Landsystem data of Bavaria}
\format{
A \code{data.frame} with 70536 observations and 3 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://dataverse.nl/dataset.xhtml?persistentId=doi:10.34894/XNC5KA}}
}
\usage{
data(landsystem_bav)
}
\description{
Landsystem data of Bavaria, Germany
}
\details{
Landsystem data of Bavaria, Germany
}
\references{
Dou, Y., Cosentino, F., Malek, Z. et al. (2021) 
A new European land systems representation accounting for landscape characteristics. Landscape Ecology. 
\url{https://doi.org/10.1007/s10980-021-01227-5}.
}
