% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{landsystem_perc_bav_tk4tel}
\alias{landsystem_perc_bav_tk4tel}
\title{Percentage cover landsystem data of Bavaria gridded to TK4tel}
\format{
A \code{data.frame} with 2267 observations and 28 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://dataverse.nl/dataset.xhtml?persistentId=doi:10.34894/XNC5KA}}
}
\usage{
data(landsystem_perc_bav)
}
\description{
Percentage cover landsystem data of Bavaria, Germany gridded to TK4tel
}
\details{
Percentage cover landsystem data of Bavaria, Germany gridded to TK4tel
}
\references{
Dou, Y., Cosentino, F., Malek, Z. et al. (2021) 
A new European land systems representation accounting for landscape characteristics. Landscape Ecology. 
\url{https://doi.org/10.1007/s10980-021-01227-5}.
}
