% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{mammals_bav}
\alias{mammals_bav}
\title{Mammal polygon data of Bavaria}
\format{A \code{data.frame} with 89 observations and 26 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://www.iucnredlist.org/resources/spatial-data-download}}
}
\usage{
data(mammals_bav)
}
\description{
Mammal polygon data of Bavaria, Germany
}
\details{
Mammal polygon data of Bavaria, Germany
}
