% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{merraclim_5m_bav_tk25}
\alias{merraclim_5m_bav_tk25}
\title{MerraClim bioclimatic data of bavaria}
\format{A \code{data.frame} with 20412 observations and 23 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.s2v81}}
}
\usage{
data(merraclim_5m_bav_tk25)
}
\description{
MerraClim bioclimatic data of Bavaria, Germany for 1980, 1990 and 2000.
}
\details{
Bioclimatic data of Bavaria, Germany for 1980, 1990 and 2000 
derived from data at 5 minute spatial resolution and resampled onto the TK25 grid.
}
\references{
C. Vega, Greta; Pertierra, Luis R.; Olalla-Tárraga, Miguel Ángel (2017), 
MERRAclim, a high-resolution global dataset of remotely sensed bioclimatic variables 
for ecological modelling, Scientific Data, https://doi.org/10.1038/sdata.2017.78
}
