% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{modis_lc_bav_tk25}
\alias{modis_lc_bav_tk25}
\title{MODIS land cover data of bavaria}
\format{A \code{data.frame} with 26975 observations and 11 variables.}
\source{
Cite this dataset as: \itemize{\item Friedl, M. A., Sulla-Menashe, D., Tan, B., Schneider, A., 
Ramankutty, N., Sibley, A., and Huang, X. (2010). MODIS Collection 5 global land cover: 
Algorithm refinements and characterization of new datasets. 
Remote Sensing of Environment, 114, 168-182. \url{https://yceo.yale.edu/modis-land-cover-product-mcd12q1}.}
}
\usage{
data(modis_lc_bav_tk25)
}
\description{
MODIS land cover data of Bavaria, Germany
}
\details{
Land cover data of Bavaria, Germany derived from MODIS satellite data and resampled onto the TK25 grid.
}
