% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{modis_lst_bav_tk25}
\alias{modis_lst_bav_tk25}
\title{MODIS land surface temperature data of bavaria}
\format{A \code{data.frame} with 2268 observations and 506 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://zenodo.org/record/1115666}}
}
\usage{
data(modis_lst_bav_tk25)
}
\description{
MODIS land surface temperature data of Bavaria, Germany
}
\details{
Monthly land surface temperature (min, max and mean) data from 2003 - 2016 
derived from MODIS satellite images, cropped by the extent of Bavaria and resampled onto the TK25 grid.
}
\references{
Metz M., Andreo V., Neteler M. (2017): A new fully gap-free time series of 
Land Surface Temperature from MODIS LST data. 
Remote Sensing, 9(12):1333. DOI: http://dx.doi.org/10.3390/rs9121333
}
