% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{odonata_bav}
\alias{odonata_bav}
\title{Odonata polygon data of Bavaria}
\format{A \code{data.frame} with 18 observations and 28 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://www.iucnredlist.org/resources/spatial-data-download}}
}
\usage{
data(odonata_bav)
}
\description{
Odonata polygon data of Bavaria, Germany
}
\details{
Odonata polygon data of Bavaria, Germany
}
