% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{odonata_tk4tel}
\alias{odonata_tk4tel}
\title{Odonata point data of Bavaria}
\format{
A \code{data.frame} with 23119 observations and 5 variables.
}
\source{
This data has been digitized from the Bavarian Atlas of Odonata.
}
\usage{
data(odonata_tk4tel)
}
\description{
Odonata point data of Bavaria, Germany
}
\details{
Odonata point data of Bavaria, Germany
}
