% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{pa_bav}
\alias{pa_bav}
\title{Protected area polygon data of Bavaria}
\format{A \code{data.frame} with 2224 observations and 29 variables.}
\source{
Cite this dataset as: \itemize{\item UNEP-WCMC and IUCN (2019), Protected Planet: 
The World Database on Protected Areas (WDPA) On-line, July/2019, Cambridge, UK: UNEP-WCMC and IUCN. 
Available at: \url{https://protectedplanet.net.}.}
}
\usage{
data(pa_bav)
}
\description{
Protected area polygon data of Bavaria, Germany
}
\details{
Protected area polygon data of Bavaria, Germany
}
