% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{pa_bav_tk25}
\alias{pa_bav_tk25}
\title{protected area data of bavaria}
\format{A \code{data.frame} with 13762 observations and 4 variables.}
\source{
Cite this dataset as: \itemize{\item UNEP-WCMC and IUCN (2019), Protected Planet: 
The World Database on Protected Areas (WDPA) On-line, July/2019, Cambridge, UK: UNEP-WCMC and IUCN. 
Available at: \url{https://protectedplanet.net.}.}
}
\usage{
data(pa_bav_tk25)
}
\description{
IUCN protected area data of Bavaria, Germany
}
\details{
Protected area data of Bavaria, Germany derived from the World Database on Protected Areas 
and resampled onto the TK25 grid.
}
