% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{plot_coords_bav}
\alias{plot_coords_bav}
\title{Plot coordinates of bavaria}
\format{A \code{data.frame} with 180 observations and 4 variables.}
\usage{
data(plot_coords_bav)
}
\description{
Plot coordinates of Bavaria, Germany
}
\details{
Plot coordinates of Bavaria, Germany.
}
