% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{predicts_deu}
\alias{predicts_deu}
\title{Predicts data of Germany}
\format{A \code{data.frame} with 97319 observations and 86 variables.}
\usage{
data(predicts_deu)
}
\description{
Predicts data of Germany
}
\details{
Predicts data of Germany. For Bavaria no data is available.
}
