% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{reptiles_bav}
\alias{reptiles_bav}
\title{Reptile polygon data of bavaria}
\format{A \code{data.frame} with 6 observations and 28 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://www.iucnredlist.org/resources/spatial-data-download}}
}
\usage{
data(reptiles_bav)
}
\description{
Reptile polygon data of Bavaria, Germany
}
\details{
IUCN reptile polygon data of Bavaria, Germany.
}
