% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{rivers_bav}
\alias{rivers_bav}
\title{River data of bavaria}
\format{A \code{data.frame} with 4556 observations and 15 variables.}
\source{
This data has been obtained from: \itemize{\item \url{https://www.hydrosheds.org/downloads}}
}
\usage{
data(rivers_bav)
}
\description{
River data of Bavaria, Germany
}
\details{
River data of Bavaria, Germany.
}
