% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{taxonomyStd}
\alias{taxonomyStd}
\title{standardised taxonomy for all species of bavaria}
\format{
A \code{data.frame} with 3745 observations and 7 variables.
}
\usage{
data(taxonomyStd)
}
\description{
Standardised taxonomy for all species of Bavaria, Germany
}
\details{
Standardised  taxonomy of Bavaria, Germany derived from all the species records 
of the LFU Biodiversity database.
}
