% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{tdm_fnf_bav}
\alias{tdm_fnf_bav}
\title{Forest/Non-forest data of bavaria}
\format{
A \code{data.frame} with 34253864 observations and 3 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://geoservice.dlr.de/web/maps/tdm:forest}}
}
\usage{
data(tdm_fnf_bav)
}
\description{
Forest/Non-forest of Bavaria, Germany
}
\details{
TanDEM-X Forest/Non-Forest data of Bavaria, Germany.
}
