% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{tk25_grid}
\alias{tk25_grid}
\title{tk25 grid data of bavaria}
\format{A \code{data.frame} with 2268 observations and 3 variables.}
\usage{
data(tk25_grid)
}
\description{
TK25 grid of Bavaria, Germany
}
\details{
TK25 grid data of Bavaria, Germany derived from the LFU Biodiversity database.
}
