% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{tk4tel_grid}
\alias{tk4tel_grid}
\title{tk4tel grid data of bavaria}
\format{
A \code{data.frame} with 621 observations and 3 variables.
}
\usage{
data(tk4tel_grid)
}
\description{
TK4tel grid of Bavaria, Germany
}
\details{
TK4tel grid data of Bavaria, Germany derived from \url{https://geo.dianacht.de/mtb/mtbshape.zip}.
}
