% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{wc1.4_2.5m_bav_tk25}
\alias{wc1.4_2.5m_bav_tk25}
\title{worldclim data of bavaria}
\format{A \code{data.frame} with 2268 observation and 69 variables.}
\source{
Cite this dataset as: \itemize{\item Hijmans, R.J., S.E. Cameron, J.L. Parra, P.G. Jones and A. Jarvis, 2005. 
Very high resolution interpolated climate surfaces for global land areas. 
International Journal of Climatology 25: 1965-1978.}
This data has been downloaded from: \itemize{\item \url{https://worldclim.org/version1}}
}
\usage{
data(wc1.4_2.5m_bav_tk25)
}
\description{
Current worldclim temperature, precipitation and bioclim data of Bavaria, Germany
}
\details{
Current temperature, precipitation and bioclimatic data of Bavaria, Germany derived from 2.5 minutes
Worldclim v1.4 and resampled onto the TK25 grid. Data is derived from interpolations of observed data, 
representative of 1960-1990.
}
