% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{wc2.0_30s_bav_tk25}
\alias{wc2.0_30s_bav_tk25}
\title{worldclim data of bavaria}
\format{A \code{data.frame} with 2268 observation and 69 variables.}
\source{
Cite this dataset as: \itemize{\item Fick, S.E. and R.J. Hijmans, 2017. Worldclim 2: 
New 1-km spatial resolution climate surfaces for global land areas. International Journal of Climatology.}
This data has been downloaded from: \itemize{\item \url{https://worldclim.org/version1}}
}
\usage{
data(wc2.0_30s_bav_tk25)
}
\description{
Current worldclim temperature, precipitation and bioclim data of Bavaria, Germany
}
\details{
Current temperature, precipitation and bioclimatic data of Bavaria, Germany derived from 30 seconds 
Worldclim v2.0 and resampled onto the TK25 grid. Data is derived from interpolations of observed data, 
representative of 1970-2000.
}
