% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{package}
\name{bdc package}
\alias{bdc package}
\alias{bdcpackage}
\title{bdc: Bavarian Data Cube}
\description{
R package to access various data sources for Bavaria, i.e. elevation, climate and land-use data.
}
\author{
RS-eco
}
\keyword{package}
