% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_clean_names.R
\name{bdc_clean_names}
\alias{bdc_clean_names}
\title{Clean scientific names}
\usage{
bdc_clean_names(sci_names)
}
\arguments{
\item{sci_names}{character. Vector of scientific names or a data frame containing the column of names to be parsed.}
}
\value{
It returns a data frame with clean and unique scientific names.
}
\description{
Clean scientific names
}
\details{
The function identifies, flags and removes suffix, infraspecific, uncertain and provisional names.
}
