% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_coordinates_from_locality.R
\name{bdc_coordinates_from_locality}
\alias{bdc_coordinates_from_locality}
\title{Identify records missing coordinates but with information on locality}
\usage{
bdc_coordinates_from_locality(
  data,
  locality = "locality",
  lon = "decimalLongitude",
  lat = "decimalLatitude"
)
}
\arguments{
\item{data}{data.frame. Containing geographical coordinates and the column
"locality'.}

\item{locality}{character string. The column with information on where records
was collected. Default = "locality".}

\item{lon}{character string. The column name with longitude. Default =
"decimalLongitude".}

\item{lat}{character string. The column name with latitude Default =
"decimalLatitude".}
}
\value{
A data.frame containing records missing coordinates but with potentially useful locality information is saved in Output/Check/01_coordinates_from_locality.csv
}
\description{
Identify records whose coordinates can potentially be extracted from
information on locality.
}
