% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_extract_cntr_names.R
\name{bdc_extract_cntr_names}
\alias{bdc_extract_cntr_names}
\title{Title: Extract_cntr_names is a function used to extracting country names in different names from wikipedia in different languages}
\usage{
bdc_extract_cntr_names(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Title: Extract_cntr_names is a function used to extracting country names in different names from wikipedia in different languages
}
