% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_filter_out_flags.R
\name{bdc_filter_out_flags}
\alias{bdc_filter_out_flags}
\title{Removes columns with the results of data quality tests}
\usage{
bdc_filter_out_flags(data, col_to_remove = "all")
}
\arguments{
\item{data}{data.frame. Containing columns to be removed}

\item{col_to_remove}{logical. Which columns should be removed? Default =
"all", which means that all columns containing the results of data quality tests
are removed.}
}
\value{
A data.frame without columns specified in 'col_to_remove'.
}
\description{
This functions is used to filter out column containing the results of data
quality tests (i.e., column starting with '.') or other columns specified.
}
\examples{
\dontrun{
data \%>\%
  bdc_flag_transposed_xy() \%>\%
  bdc_filter_out_flags()
}
}
