% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_flag_missing_xy.R
\name{bdc_flag_missing_xy}
\alias{bdc_flag_missing_xy}
\title{Flag missing coordinates}
\usage{
bdc_flag_missing_xy(data, lon, lat)
}
\arguments{
\item{data}{a data.frame with the default column names: "database_id",
"scientificName", "decimalLongitude", "decimalLatitude"}
}
\description{
This function add a new column \code{.invalid_xy} in the returned dataset
}
\examples{
\dontrun{
data \%>\%
  bdc_flag_invalid_xy()
}
}
