% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_flag_xy_provenace.R
\name{bdc_flag_xy_provenance}
\alias{bdc_flag_xy_provenance}
\title{Title Flag records from doubtful provenance (e.g., 'fossil', 'HUCP') whose can be unreliable or unsuitable for certain analyses}
\usage{
bdc_flag_xy_provenance(data, basisOfRecord)
}
\arguments{
\item{basisOfRecord}{character string. The column with basis of records}

\item{data.}{data.frame. Containing basis of records column.}
}
\value{

}
\description{
Title Flag records from doubtful provenance (e.g., 'fossil', 'HUCP') whose can be unreliable or unsuitable for certain analyses
}
