% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_geographic_outlier.R
\name{bdc_geographic_outlier}
\alias{bdc_geographic_outlier}
\title{Identify plant species records outside their known geographic distribution}
\usage{
bdc_geographic_outlier(
  data,
  sci_names = "scientificName",
  lat = "decimalLatitude",
  lon = "decimalLongitude"
)
}
\arguments{
\item{data}{data.frame. A data frame containing geographic coordinates and
scientific names.}

\item{sci_names}{character string. The column name with species scientific
name. Default = "scientificName".}

\item{lat}{character string. The column name with latitude. Coordinates must
be expressed in decimal degree and in WGS84. Default = "decimalLatitude".}

\item{lon}{character string. The column with longitude. Coordinates must be
expressed in decimal degree and in WGS84. Default = "decimalLongitude".}
}
\value{
A data.frame contain the column ".otl". Compliant
(TRUE) if 'lat' and 'lon' are not empty; otherwise "FALSE".
}
\description{
This function flags records outside their known geographical distribution
(i.e., Brazilian states). Geographical distribution information is obtained
from Flora do Brasil 2020.
}
\details{
Check if coordinates are valid before run this test.
Species occurrence records in Goias and range in DF are assigned as TRUE.
Similarly, species occurrence in DF and range is GO are considered TRUE.
}
\examples{
\dontrun{
}
}
