% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_geographic_outlier.R
\name{bdc_geographic_outlier}
\alias{bdc_geographic_outlier}
\title{Identify plant species records outside known geographic distribution}
\usage{
bdc_geographic_outlier(
  data,
  species = "scientificName",
  longitude = "decimalLongitude",
  latitude = "decimalLatitude"
)
}
\arguments{
\item{data}{data.frame. A data frame containing geographic coordinates and species names.}

\item{species}{character string. The column with scientific names.}

\item{longitude}{character string. The column with longitude in decimal degrees. Default = "decimalLongitude"}

\item{latitude}{character string. The column with the latitude in decimal degrees. Default = "decimalLatitude".}
}
\value{
flag of records are outliers (FALSE) or not (TRUE)
}
\description{
Flags geographic outliers by comparing the status where record is against
species range (i.e., Brazilian states) obtained from the Flora do Brazil
2020.
}
\details{
Check if coordinates are valid before run this test.
Species occurrence records in Goias and range in DF are assigned as TRUE.
Similarly, species occurrence in DF and range is GO are considered TRUE.
}
\examples{
\dontrun{
}
}
