% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{bdc_get_taxa_taxadb}
\alias{bdc_get_taxa_taxadb}
\title{Title: Get names using taxadb R package. Fuzzy match is allowed by using a modification version of taxadb get_filter names function}
\usage{
bdc_get_taxa_taxadb(
  taxa,
  replace.synonyms = TRUE,
  suggest.names = TRUE,
  suggestion.distance = 0.9,
  parse = FALSE,
  db = NULL
)
}
\arguments{
\item{taxa}{}

\item{replace.synonyms}{}

\item{suggest.names}{}

\item{suggestion.distance}{}

\item{parse}{}

\item{db}{}
}
\value{

}
\description{
Title: Get names using taxadb R package. Fuzzy match is allowed by using a modification version of taxadb get_filter names function
}
