% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_get_wiki_country.R
\name{bdc_get_wiki_country}
\alias{bdc_get_wiki_country}
\title{Title: Countries names in different language}
\usage{
bdc_get_wiki_country()
}
\value{

}
\description{
Title: Countries names in different language
}
