% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_get_wiki_country.R
\name{bdc_get_wiki_country}
\alias{bdc_get_wiki_country}
\title{Internal function. Download countries names in different language from
wikipedia}
\usage{
bdc_get_wiki_country()
}
\value{
A data.frame with country names in different languages. This database
is also saved in 'data/countries_names.csv'.
}
\description{
This is a helper function for downloading country names in different
languages from Wikipedia. A CSV file is exported to 'data/countries_names'.
}
