% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_query_names_taxadb.R
\name{bdc_query_names_taxadb}
\alias{bdc_query_names_taxadb}
\title{Matching scientific names against local stored taxonomic authority database}
\usage{
bdc_query_names_taxadb(
  sci_name,
  replace_synonyms = TRUE,
  suggest_names = TRUE,
  suggestion_distance = 0.9,
  db = "gbif",
  rank_name = NULL,
  rank = NULL,
  parallel = FALSE,
  ncores = 2,
  export_accepted = FALSE
)
}
\arguments{
\item{sci_name}{A character vector of species names. The function does not
clean species names (e.g.: infraespecific, var., inf.), it is expected
clean names. The inclusion of 'var.' increases name distances and it is
advised to set smaller suggestion_distance values.}

\item{replace_synonyms}{A logical value (default = TRUE) whether synonyms
must be replaced by the valid names found in taxadb database.}

\item{suggest_names}{A logical value (default = TRUE) whether species names
must be suggested if it is not found in the first search.}

\item{suggestion_distance}{A numeric value (default = 0.9). It is the
accepted distance between the searched names and suggested ones. Distances
higher than specified here are not suggested and NA is returned.}

\item{db}{Character string. Contains the name of the taxonomic database where
information should be searched. By default, 'gbif' will be installed. Use
'all' to install all available taxonomic database automatically.}

\item{rank_name}{character string. A taxonomic rank used to be used to filter
data. Options available are: "kingdom", "phylum", "class", "order",
"family", and "genus".}

\item{rank}{character string. Taxonomic rank name (e.g. "Plantae",
"Animalia", "Aves", "Carnivora". Default is NULL.}

\item{parallel}{A logical value indicating whether distance calculation
should be done in parallel.}

\item{ncores}{Number of cores to run in parallel.}

\item{export_accepted}{A logical value (default = FALSE) whether a table is
exported containing all species with more than one valid name to be further
explored by the user.}
}
\value{
This function returns a data.frame with the same number of rows and
order than sci_name with the information provided by the database.
}
\description{
This function borrows heavily from the great work done by Norman et al.
(2020) in their package 'taxadb', which contains functions that allow
querying millions of taxonomic names in a fast and consistent way. The
bdc_get_taxadb contains additions that allow queries using fuzzy matching and
converting synonyms to the current accepted name. The fuzzy match algorithm
was inspired by get.taxa function of the 'flora' package.
}
\examples{
\dontrun{
sci_name <- c(
  "Polystachya estrellensis",
  "Tachigali rubiginosa",
  "Oxalis rhombeo ovata",
  "Axonopus canescens",
  "Prosopis",
  "Guapira opposita",
  "Clidemia naevula",
  "Poincianella pyramidalis",
  "Hymenophyllum polyanthos")

names_harmonization <-
  bdc_query_names_taxadb(
  sci_names,
  replace_synonyms = TRUE,
  suggest_names = TRUE,
  suggestion_distance = 0.9,
  db = "gbif",
  parallel = FALSE,
  ncores = 2,
  export_accepted = FALSE)
}
}
