% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_quickmap.R
\name{bdc_quickmap}
\alias{bdc_quickmap}
\title{Title: Create a map of points using ggplot2}
\usage{
bdc_quickmap(data, lon, lat, column_to_map)
}
\arguments{
\item{data}{}

\item{lon}{}

\item{lat}{}
}
\value{

}
\description{
Title: Create a map of points using ggplot2
}
