% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{bdc_quickmap}
\alias{bdc_quickmap}
\title{Title: Create a map of points using ggplot2}
\usage{
bdc_quickmap(data, long, lat)
}
\arguments{
\item{data}{}

\item{lat}{}

\item{lon}{}
}
\value{

}
\description{
Title: Create a map of points using ggplot2
}
