% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_round_dec.R
\name{bdc_round_dec}
\alias{bdc_round_dec}
\title{Title: Function to test round coordinates}
\usage{
bdc_round_dec(
  data,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  ndec = c(0, 1, 2)
)
}
\arguments{
\item{data:}{data.frame. A data.frame with coordinates data}

\item{lon:}{character. Column names with longitude values}

\item{lat:}{character. Column names with latitude values}

\item{ndec:}{ndec: numeric. A vector with number of decimal to be tested. Default ndec=c(0,1,2)}
}
\value{

}
\description{
Title: Function to test round coordinates
}
