% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_standardize_country.R
\name{bdc_standardize_country}
\alias{bdc_standardize_country}
\title{Title: standard_country is a function to correct, standardize, and assign a ISO code to country names}
\usage{
bdc_standardize_country(data, country, country_names_db)
}
\arguments{
\item{data}{}

\item{country}{}

\item{country_names_db}{}
}
\value{

}
\description{
Title: standard_country is a function to correct, standardize, and assign a ISO code to country names
}
