% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_standardize_datasets.R
\name{bdc_standardize_datasets}
\alias{bdc_standardize_datasets}
\title{Title: Standardize datasets columns based on metadata file}
\usage{
bdc_standardize_datasets(metadata)
}
\arguments{
\item{metadata}{a table containing information about which columns of the
original dataset need to be renamed following Darwin Core terminology.
Please see the \code{Config/DatabaseInfo.csv} file.}
}
\description{
Title: Standardize datasets columns based on metadata file
}
