% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc_suggest_names_taxadb.R
\name{bdc_suggest_names_taxadb}
\alias{bdc_suggest_names_taxadb}
\title{Title: Function used to standardized taxonomic names using taxadb R packpage. This functions is a modification of get.taxa function (flora package) inserted in get_names function (taxadb package) for allowing fuzzy matching.}
\usage{
bdc_suggest_names_taxadb(sci_name, max.distance = 0.75, provider)
}
\arguments{
\item{sci_name}{}

\item{max.distance}{}

\item{provider}{}
}
\value{

}
\description{
Title: Function used to standardized taxonomic names using taxadb R packpage. This functions is a modification of get.taxa function (flora package) inserted in get_names function (taxadb package) for allowing fuzzy matching.
}
